# V2.0 speedup rpn