--- a/mmdet/ops/__init__.py
+++ b/mmdet/ops/__init__.py
@@ -8,7 +8,7 @@ from .dcn import (DeformConv, DeformConvPack, DeformRoIPooling,
                   deform_conv, deform_roi_pooling, modulated_deform_conv)
 from .generalized_attention import GeneralizedAttention
 from .masked_conv import MaskedConv2d
-from .nms import nms, soft_nms
+from .nms import batched_nms, nms, soft_nms
 from .non_local import NonLocal2D
 from .norm import build_norm_layer
 from .plugin import build_plugin_layer
@@ -28,5 +28,5 @@ __all__ = [
     'MaskedConv2d', 'ContextBlock', 'GeneralizedAttention', 'NonLocal2D',
     'get_compiler_version', 'get_compiling_cuda_version', 'build_conv_layer',
     'ConvModule', 'ConvWS2d', 'conv_ws_2d', 'build_norm_layer', 'Scale',
-    'build_upsample_layer', 'build_plugin_layer'
+    'build_upsample_layer', 'build_plugin_layer', 'batched_nms'
 ]