--- a/mmdet/ops/nms/__init__.py
+++ b/mmdet/ops/nms/__init__.py
@@ -1,3 +1,3 @@
-from .nms_wrapper import nms, soft_nms
+from .nms_wrapper import batched_nms, nms, soft_nms
 
-__all__ = ['nms', 'soft_nms']
+__all__ = ['nms', 'soft_nms', 'batched_nms']