--- a/mmdet/ops/nms/nms_wrapper.py
+++ b/mmdet/ops/nms/nms_wrapper.py
@@ -116,3 +116,23 @@ def soft_nms(dets, iou_thr, method='linear', sigma=0.5, min_score=1e-3):
     else:
         return new_dets.numpy().astype(dets.dtype), inds.numpy().astype(
             np.int64)
+
+
+def batched_nms(bboxes, scores, idxs, nms_cfg):
+    # Modified from https://github.com/pytorch/vision/blob
+    # /505cd6957711af790211896d32b40291bea1bc21/torchvision/ops/boxes.py#L39.
+    # strategy: in order to perform NMS independently per class.
+    # we add an offset to all the boxes. The offset is dependent
+    # only on the class idx, and is large enough so that boxes
+    # from different classes do not overlap
+    max_coordinate = bboxes.max()
+    offsets = idxs.to(bboxes) * (max_coordinate + 1)
+    bboxes_for_nms = bboxes + offsets[:, None]
+    nms_cfg_ = nms_cfg.copy()
+    nms_type = nms_cfg_.pop('type', 'nms')
+    nms_op = eval(nms_type)
+    dets, keep = nms_op(
+        torch.cat([bboxes_for_nms, scores[:, None]], -1), **nms_cfg_)
+    bboxes = bboxes[keep]
+    scores = dets[:, -1]
+    return torch.cat([bboxes, scores[:, None]], -1), keep