--- a/mmdet/ops/nms/src/cuda/nms_kernel.cu
+++ b/mmdet/ops/nms/src/cuda/nms_kernel.cu
@@ -132,8 +132,7 @@ at::Tensor nms_cuda_forward(const at::Tensor boxes, float nms_overlap_thresh) {
 
   THCudaFree(state, mask_dev);
   // TODO improve this part
-  return std::get<0>(order_t.index({
-                       keep.narrow(/*dim=*/0, /*start=*/0, /*length=*/num_to_keep).to(
-                         order_t.device(), keep.scalar_type())
-                     }).sort(0, false));
+  return order_t.index({
+      keep.narrow(/*dim=*/0, /*start=*/0, /*length=*/num_to_keep).to(
+          order_t.device(), keep.scalar_type())});
 }