Speed up RPN head.

| Model             | Original Inf time (s/iter) | New Inf time (s/iter) |
|-------------------|----------------------------|-----------------------|
| Faster R-CNN R-50 | 16.0                       | 18.0                  |
| Mask R-CNN R-50   | 10.9                       | 11.9                  |
| Retinanet R-50    | 16                         |   16.1                    |